+++ src/lib.rs
@@ -5,8 +5,8 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
 use bevy::{
-    prelude::*,
     platform::collections::{HashMap, HashSet},
+    prelude::*,
 };
 use bevy_image::{Image, ImageSampler};
 use derive_setters::Setters;